Fix: Avoid using $this within the included file.
By using a closure it is possible to remove the context of the class and avoid
to use the $this ref within the template file.

Also, since the $this is allowed within the closure and functions we don't use
a classic anonymous function but a closure bind to null.

Added test too