Fix: Plugin dir path return wrong path if used as composer package.
When the templateloader is used as a composer package the `plugin_dir_path`
doesn't return the expected value due to the parameter `__DIR__` that reflect
the path where the `Plugin` class is located.

Means the correct plugin dir path will never reached even if templateloader
is used into a Wordpress plugin.

So, remove the `Plugin` class and introduce an additional argument to the
`Loader` construct that allow us to define a `defaultPath` where search
if any template file is found within the theme/child.

Also, remove `locateFile` from the interface because it's must be the
business of the class. The TemplateLoader load template not locate files.

fix #6